+++ apps/docs/src/content/02-foundations/03-content-guidelines/02-informationskonzept/examples/alertBadge.tsx
@@ -12,4 +12,8 @@ import { AlertBadge } from "@mittwald/flow-react-components";
   </AlertBadge>
 
   <AlertBadge status="success">Verifiziert</AlertBadge>
+
+  <AlertBadge status="unavailable">
+    Nicht gefunden
+  </AlertBadge>
 </Row>;