+++ apps/docs/src/content/04-components/content/color/examples/status.tsx
@@ -5,4 +5,5 @@ import { Color } from "@mittwald/flow-react-components";
   <Color color="warning">warning</Color>
   <Color color="info">info</Color>
   <Color color="success">success</Color>
+  <Color color="unavialable">unavialable</Color>
 </Row>;