+++ packages/components/src/components/Color/stories/Default.stories.tsx
@@ -4,6 +4,7 @@ import { Text } from "@/components/Text";
 import { Color } from "@/components/Color";
 import Heading from "@/components/Heading";
 import { Section } from "@/components/Section";
+import { statusTypes } from "@/lib/types/props";
 
 const meta: Meta<typeof Color> = {
   title: "Content/Color",
@@ -12,16 +13,7 @@ const meta: Meta<typeof Color> = {
   argTypes: {
     color: {
       control: "inline-radio",
-      options: [
-        "blue",
-        "violet",
-        "teal",
-        "lilac",
-        "danger",
-        "warning",
-        "info",
-        "success",
-      ],
+      options: ["blue", "violet", "teal", "lilac", ...statusTypes],
     },
   },
   render: (props) => (