+++ packages/components/src/components/List/stories/Default.stories.tsx
@@ -74,6 +74,7 @@ const meta: Meta<typeof List> = {
           <DomainList.List
             batchSize={5}
             aria-label="Domains"
+            onAction={(domain) => console.log(domain.hostname)}
             settingStorageKey="domains"
             defaultViewMode="tiles"
           >
@@ -142,11 +143,7 @@ const meta: Meta<typeof List> = {
               </DomainList.TableBody>
             </DomainList.Table>
 
-            <DomainList.Item
-              href={() => "#"}
-              showTiles
-              textValue={(domain) => domain.hostname}
-            >
+            <DomainList.Item showTiles textValue={(domain) => domain.hostname}>
               {(domain) => (
                 <ListItemView>
                   <Avatar color={domain.type === "Domain" ? "blue" : "teal"}>